+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `Innmind\Json\Json::encode()` and `::decode()` are now declared pure
+- `Innmind\Json\Json::maybeDecode()`
 
 ### Removed